refactored 'database' directory into assets and changed relevant routes
